requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
whisperx==3.1.1
redis==5.0.1
paho-mqtt==2.1.0
pydantic==2.4.2
pydantic-settings==2.8.1
python-jose==3.3.0
passlib==1.7.4
pytest==7.4.2
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
bcrypt==4.0.1
SQLAlchemy==2.0.21
alembic==1.12.0
